FIX: 2/30 LUT: int -> float when parsing bin boundaries and replacement values

+++ 2_processInFiji/30_LUT.py
@@ -30,7 +30,7 @@ def parseBinBoundaries(valString):
 		return outVals
 
 	while a < b:
-		v = int(valString[a:b])
+		v = float(valString[a:b])
 		outVals.append(v)
 
 		a = b+1
@@ -48,7 +48,7 @@ def parseBinValues(valString):
 		b = len(valString)
 
 	while a < b:
-		v = int(valString[a:b])
+		v = float(valString[a:b])
 		outVals.append(v)
 
 		a = b+1
